+++ src/types/preview.d.ts
@@ -1,0 +1,7 @@
+import { ID } from "./siyuan";
+
+export interface IPreview {
+    display: boolean; // 是否显示
+    id: ID; // 块 ID
+    focus: boolean; // 是否聚焦
+}